chore: update version to 1.6.0, refactor API routes by renaming modules, and remove deprecated admin and function APIs for improved structure and maintainability
