+++ _public/static/admin/pages/login.html
@@ -24,8 +24,8 @@
       }
     }
   </script>
-  <link href="/static/common/css/common.css?v=1.5.7" rel="stylesheet">
-  <link href="/static/common/css/login.css?v=1.5.7" rel="stylesheet">
+  <link href="/static/common/css/common.css?v=1.6.0" rel="stylesheet">
+  <link href="/static/common/css/login.css?v=1.6.0" rel="stylesheet">
   <link href="/static/common/css/toast.css" rel="stylesheet">
 </head>
 
@@ -56,10 +56,10 @@
   <div id="app-footer"></div>
 
   <script src="/static/i18n/i18n.js"></script>
-  <script src="/static/common/js/admin-auth.js?v=1.5.7"></script>
+  <script src="/static/common/js/admin-auth.js?v=1.6.0"></script>
   <script src="/static/common/js/toast.js"></script>
-  <script src="/static/common/js/footer.js?v=1.5.7"></script>
-  <script src="/static/admin/js/login.js?v=1.5.7"></script>
+  <script src="/static/common/js/footer.js?v=1.6.0"></script>
+  <script src="/static/admin/js/login.js?v=1.6.0"></script>
 </body>
 
 </html>
