+++ _public/static/common/js/footer.js
@@ -2,7 +2,7 @@ async function loadAdminFooter() {
   const container = document.getElementById('app-footer');
   if (!container) return;
   try {
-    const res = await fetch('/static/common/html/footer.html?v=1.5.7');
+    const res = await fetch('/static/common/html/footer.html?v=1.6.0');
     if (!res.ok) return;
     container.innerHTML = await res.text();
   } catch (e) {
