+++ _public/static/common/js/function-header.js
@@ -2,7 +2,7 @@ async function loadFunctionHeader() {
   const container = document.getElementById('app-header');
   if (!container) return;
   try {
-    const res = await fetch('/static/common/html/function-header.html?v=1.5.4');
+    const res = await fetch('/static/common/html/function-header.html?v=1.6.0');
     if (!res.ok) return;
     container.innerHTML = await res.text();
     const logoutBtn = container.querySelector('#function-logout-btn');
