+++ _public/static/common/js/header.js
@@ -2,7 +2,7 @@ async function loadAdminHeader() {
   const container = document.getElementById('app-header');
   if (!container) return;
   try {
-    const res = await fetch('/static/common/html/header.html?v=1.5.7');
+    const res = await fetch('/static/common/html/header.html?v=1.6.0');
     if (!res.ok) return;
     container.innerHTML = await res.text();
     const path = window.location.pathname;
