+++ _public/static/function/pages/imagine.html
@@ -9,8 +9,8 @@
   <script src="https://cdn.tailwindcss.com"></script>
   <link href="https://cdn.jsdelivr.net/npm/geist@1.0.0/dist/fonts/geist-sans/style.css" rel="stylesheet">
   <link href="https://cdn.jsdelivr.net/npm/geist@1.0.0/dist/fonts/geist-mono/style.css" rel="stylesheet">
-  <link href="/static/common/css/common.css?v=1.5.4" rel="stylesheet">
-  <link href="/static/function/css/imagine.css?v=1.5.4" rel="stylesheet">
+  <link href="/static/common/css/common.css?v=1.6.0" rel="stylesheet">
+  <link href="/static/function/css/imagine.css?v=1.6.0" rel="stylesheet">
   <link href="/static/common/css/toast.css" rel="stylesheet">
 </head>
 
@@ -238,12 +238,12 @@
   <div id="app-footer"></div>
 
   <script src="/static/i18n/i18n.js"></script>
-  <script src="/static/common/js/admin-auth.js?v=1.5.4"></script>
-  <script src="/static/common/js/function-header.js?v=1.5.4"></script>
-  <script src="/static/common/js/footer.js?v=1.5.4"></script>
+  <script src="/static/common/js/admin-auth.js?v=1.6.0"></script>
+  <script src="/static/common/js/function-header.js?v=1.6.0"></script>
+  <script src="/static/common/js/footer.js?v=1.6.0"></script>
   <script src="/static/common/js/toast.js"></script>
   <script src="https://cdnjs.cloudflare.com/ajax/libs/jszip/3.10.1/jszip.min.js"></script>
-  <script src="/static/function/js/imagine.js?v=1.5.4"></script>
+  <script src="/static/function/js/imagine.js?v=1.6.0"></script>
 </body>
 
 </html>
