+++ _public/static/function/pages/voice.html
@@ -9,8 +9,8 @@
   <script src="https://cdn.tailwindcss.com"></script>
   <link href="https://cdn.jsdelivr.net/npm/geist@1.0.0/dist/fonts/geist-sans/style.css" rel="stylesheet">
   <link href="https://cdn.jsdelivr.net/npm/geist@1.0.0/dist/fonts/geist-mono/style.css" rel="stylesheet">
-  <link href="/static/common/css/common.css?v=1.5.4" rel="stylesheet">
-  <link href="/static/function/css/voice.css?v=1.5.4" rel="stylesheet">
+  <link href="/static/common/css/common.css?v=1.6.0" rel="stylesheet">
+  <link href="/static/function/css/voice.css?v=1.6.0" rel="stylesheet">
   <link href="/static/common/css/toast.css" rel="stylesheet">
   <script src="https://cdn.jsdelivr.net/npm/livekit-client@2.7.3/dist/livekit-client.umd.min.js"></script>
 </head>
@@ -148,11 +148,11 @@
   <div id="app-footer"></div>
 
   <script src="/static/i18n/i18n.js"></script>
-  <script src="/static/common/js/admin-auth.js?v=1.5.4"></script>
-  <script src="/static/common/js/function-header.js?v=1.5.4"></script>
-  <script src="/static/common/js/footer.js?v=1.5.4"></script>
+  <script src="/static/common/js/admin-auth.js?v=1.6.0"></script>
+  <script src="/static/common/js/function-header.js?v=1.6.0"></script>
+  <script src="/static/common/js/footer.js?v=1.6.0"></script>
   <script src="/static/common/js/toast.js"></script>
-  <script src="/static/function/js/voice.js?v=1.5.4"></script>
+  <script src="/static/function/js/voice.js?v=1.6.0"></script>
 </body>
 
 </html>
