+++ app/api/v1/admin/__init__.py
@@ -2,9 +2,9 @@
 
 from fastapi import APIRouter
 
-from app.api.v1.admin_api.cache import router as cache_router
-from app.api.v1.admin_api.config import router as config_router
-from app.api.v1.admin_api.token import router as tokens_router
+from app.api.v1.admin.cache import router as cache_router
+from app.api.v1.admin.config import router as config_router
+from app.api.v1.admin.token import router as tokens_router
 
 router = APIRouter()
