+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "grok2api"
-version = "1.5.7"
+version = "1.6.0"
 description = "Grok2API rebuilt with FastAPI, fully aligned with the latest web call format. Supports streaming and non-streaming chat, image generation/editing, deep thinking, token pool concurrency, and automatic load balancing."
 readme = "README.md"
 requires-python = ">=3.13"
